Allow for who list level to be stored as a uint

+++ src/server/game/Storages/WhoListStorage.h
@@ -24,7 +24,7 @@
 class WhoListPlayerInfo
 {
 public:
-    WhoListPlayerInfo(ObjectGuid guid, uint32 team, AccountTypes security, uint8 level, uint8 clss, uint8 race, uint32 zoneid, uint8 gender, bool visible, std::wstring const& widePlayerName,
+    WhoListPlayerInfo(ObjectGuid guid, uint32 team, AccountTypes security, uint32 level, uint8 clss, uint8 race, uint32 zoneid, uint8 gender, bool visible, std::wstring const& widePlayerName,
         std::wstring const& wideGuildName, std::string const& playerName, std::string const& guildName) :
         _guid(guid), _team(team), _security(security), _level(level), _class(clss), _race(race), _zoneid(zoneid), _gender(gender), _visible(visible),
         _widePlayerName(widePlayerName), _wideGuildName(wideGuildName), _playerName(playerName), _guildName(guildName) {}
@@ -32,7 +32,7 @@ public:
     ObjectGuid GetGuid() const { return _guid; }
     uint32 GetTeam() const { return _team; }
     AccountTypes GetSecurity() const { return _security; }
-    uint8 GetLevel() const { return _level; }
+    uint32 GetLevel() const { return _level; }
     uint8 GetClass() const { return _class; }
     uint8 GetRace() const { return _race; }
     uint32 GetZoneId() const { return _zoneid; }
@@ -47,7 +47,7 @@ private:
     ObjectGuid _guid;
     uint32 _team;
     AccountTypes _security;
-    uint8 _level;
+    uint32 _level;
     uint8 _class;
     uint8 _race;
     uint32 _zoneid;
